+++ ajax/inputtext.php
@@ -30,6 +30,8 @@
  * ---------------------------------------------------------------------
  */
 
+use Glpi\Toolbox\Sanitizer;
+
 $AJAX_INCLUDE = 1;
 
 include ('../inc/includes.php');
@@ -43,6 +45,6 @@ if (isset($_POST['name'])) {
    echo "<input type='text' ".(isset($_POST["size"])?" size='".$_POST["size"]."' ":"")." ".
          (isset($_POST["maxlength"])?"maxlength='".$_POST["maxlength"]."' ":"")." name='".
          $_POST['name']."' value=\"".
-         Html::cleanInputText(Toolbox::clean_cross_side_scripting_deep(rawurldecode(stripslashes($_POST["data"])))).
+         Html::cleanInputText(Sanitizer::sanitize(rawurldecode(stripslashes($_POST["data"])))).
         "\">";
 }